Sample apps fails to connect to TLS enabled clusters that don't need a client cert (#52)
sslmode is set to disabled by default. TLS can be enabled for postgres without a client cert, so it seems that we could use the default sslmode which is 'prefer' instead of disabling it outright.

Test Plan: Run against a cluster with TLS enabled but no client sort, verify connectivity works.