Integrated the jdbc smart driver with Sql* sample apps. (#69)
* With this change now the yb-sample apps Sqlf* workloads will can use
both PostgreSQL jdbc driver and Yugabyte smart jdbc driver depending on
the value of load-balance argument value. By default the load-balance
will be true which means it will use the smart driver. Only when the
value of load-balance=false ( specified as part of startup argument )
the PostgreSQL driver will be used.